First Phase testing
Added a DEBUG print method to Relation. Partition seems to be working (at least for the example in the assignment).
Need to figure out how to pick H1_N: the amount of bits taken into account from H1 hash function, as it also affects the number of buckets (which should not be more than the size of the Relation)